requirements.txt
asgiref==1.1.2
asn1crypto==0.22.0
attrs==17.2.0
autobahn==17.7.1
Automat==0.6.0
cached-property==1.3.0
cffi==1.10.0
channels==1.1.6
constantly==15.1.0
cryptography==2.0.2
daphne==1.3.0
Django==1.11.3
django-debug-toolbar==1.8
django-extensions==1.8.1
django-filter==1.0.4
django-url-filter==0.3.3
djangorestframework==3.6.3
enum34==1.1.6
gunicorn==19.7.1
h2==3.0.1
hpack==3.0.0
hyperframe==5.1.0
hyperlink==17.3.0
idna==2.5
incremental==17.5.0
Markdown==2.6.8
markdown2==2.3.4
priority==1.3.0
psycopg2==2.7.1
pyasn1==0.3.1
pyasn1-modules==0.0.10
pycparser==2.18
pyOpenSSL==17.2.0
pytz==2017.2
service-identity==17.0.0
six==1.10.0
sqlparse==0.2.3
Twisted==17.5.0
txaio==2.8.1
whitenoise==3.3.0
zope.interface==4.4.2
